added category route in categorical bar graph
