+++ client/components/ChartComponents/assets/categoryColors.js
@@ -1,0 +1,17 @@
+const barColors = [
+  "#D99694",
+  "#F87D79",
+  '#FAC090',
+  "#FFFD70",
+  "#C3D69B",
+  "#78ED8C",
+  "#00FFFF",
+  "#93CDDD",
+  "#558ED5",
+  "#9675F9",
+  "#FA75F8",
+  "#E5B2FF",
+  "#CEAE8E"
+];
+
+export default barColors;
